requirements.txt
Django>=3.2.7
django-extensions
django-filter
djangorestframework
markdown
psycopg2-binary>=2.9.1
